big merge of 2388redis & 2669tor (#2790)

+++ home.admin/_provision.setup.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# check if started with sudo
+if [ "$EUID" -ne 0 ]; then 
+  echo "error='run as root'"
+  exit 1
+fi
+
 # this provision file is just executed on fresh setups
 # not on recoveries or updates
 
@@ -23,35 +29,29 @@ echo "" > ${logFile}
 echo "###################################" >> ${logFile}
 echo "# _provision.setup.sh" >> ${logFile}
 echo "###################################" >> ${logFile}
-sudo sed -i "s/^message=.*/message='Provision Setup'/g" ${infoFile}
 
 ###################################
 # Preserve SSH keys
 # just copy dont link anymore
 # see: https://github.com/rootzoll/raspiblitz/issues/1798
-sed -i "s/^message=.*/message='SSH Keys'/g" ${infoFile}
+/home/admin/_cache.sh set message "SSH Keys"
 
 # link ssh directory from SD card to HDD
 /home/admin/config.scripts/blitz.ssh.sh backup
 
 ###################################
 # Prepare Blockchain Service
-sed -i "s/^message=.*/message='Blockchain Setup'/g" ${infoFile}
+/home/admin/_cache.sh set message "Blockchain Setup"
+source <(/home/admin/_cache.sh get network chain hddBlocksBitcoin)
 
 if [ "${network}" == "" ]; then
-  sed -i "s/^state=.*/state=error/g" ${infoFile}
-  sed -i "s/^message=.*/message='config: missing network'/g" ${infoFile}
-  echo "FAIL see ${logFile}"
-  echo "FAIL: missing network in (${setupFile})!" >> ${logFile}
-  exit 20
+  /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "missing-network" "" "" ${logFile}
+  exit 2
 fi
 
 if [ "${chain}" == "" ]; then
-  sed -i "s/^state=.*/state=error/g" ${infoFile}
-  sed -i "s/^message=.*/message='config: missing chain'/g" ${infoFile}
-  echo "FAIL see ${logFile}"
-  echo "FAIL: missing chain in (${setupFile})!" >> ${logFile}
-  exit 2
+  /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "missing-chain" "" "" ${logFile}
+  exit 3
 fi
 
 # copy configs files and directories
@@ -74,7 +74,7 @@ chown -R admin:admin /home/admin/.${network} >>${logFile} 2>&1
 confExists=$(sudo ls /mnt/hdd/${network}/${network}.conf | grep -c "${network}.conf")
 echo "File Exists: /mnt/hdd/${network}/${network}.conf --> ${confExists}" >> ${logFile}
 
-# set password B as RPC password
+# set password B as RPC password (from setup file)
 echo "# setting PASSWORD B" >> ${logFile}
 /home/admin/config.scripts/blitz.setpassword.sh b "${passwordB}" >> ${logFile}
 
@@ -119,9 +119,7 @@ do
   sync
   loopcount=$(($loopcount +1))
   if [ ${loopcount} -gt 50 ]; then
-    sed -i "s/^state=.*/state=error/g" ${infoFile}
-    sed -i "s/^message=.*/message='setup: failed ${network}'/g" ${infoFile}
-    echo "FAIL: setup: failed ${network}" >> ${logFile}
+    /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "btc-testrun-fail" "${network}d not running" "sudo -u bitcoin ${network}-cli getblockchaininfo | grep "initialblockdownload" -c --> ${bitcoinRunning}" ${logFile}
     exit 4
   fi
 done
@@ -130,14 +128,20 @@ echo "OK ${network} startup successful " >> ${logFile}
 
 ###################################
 # Prepare Lightning
+source /mnt/hdd/raspiblitz.conf
 echo "Prepare Lightning (${lightning})" >> ${logFile}
+
+if [ "${hostname}" == "" ]; then
+  /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "missing-hostname" "" "" ${logFile}
+  exit 41
+fi
 
 if [ "${lightning}" != "lnd" ]; then 
 
   ###################################
   # Remove LND from systemd
   echo "Remove LND" >> ${logFile}
-  sudo sed -i "s/^message=.*/message='Deactivate Lightning'/g" ${infoFile}
+  /home/admin/_cache.sh set message "Deactivate Lightning"
   sudo systemctl disable lnd
   sudo rm /etc/systemd/system/lnd.service 2>/dev/null
   sudo systemctl daemon-reload
@@ -148,30 +152,24 @@ if [ "${lightning}" == "lnd" ]; then
   ###################################
   # LND
   echo "############## Setup LND" >> ${logFile}
-  sudo sed -i "s/^message=.*/message='LND Setup'/g" ${infoFile}
+  /home/admin/_cache.sh set message "LND Setup"
 
+  # password C (raspiblitz.setup)
   if [ "${passwordC}" == "" ]; then
-    sed -i "s/^state=.*/state=error/g" ${infoFile}
-    sed -i "s/^message=.*/message='config: missing passwordC'/g" ${infoFile}
-    echo "FAIL see ${logFile}"
-    echo "FAIL: missing passwordC in (${setupFile})!" >> ${logFile}
+    /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "missing-passwordc" "config: missing passwordC" "" ${logFile}
     exit 5
   fi
 
-  # if user uploaded an LND rescue file
+  # if user uploaded an LND rescue file (raspiblitz.setup)
   if [ "${lndrescue}" != "" ]; then
     echo "Restore LND data from uploaded rescue file ${lndrescue} ..." >> ${logFile}
     source <(sudo /home/admin/config.scripts/lnd.backup.sh lnd-import ${lndrescue})
     if [ "${error}" != "" ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: lnd import backup failed'/g" ${infoFile}
-      echo "FAIL see ${logFile}"
-      echo "FAIL: setup: lnd import backup failed" >> ${logFile}
-      echo "${error}" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lndrescue-import" "setup: lnd import backup failed" "${error}" ${logFile}
       exit 6
     fi
   else
-    # preparing new LND config
+    # preparing new LND config (raspiblitz.setup)
     echo "Creating new LND config ..." >> ${logFile}
     sudo -u bitcoin mkdir /mnt/hdd/lnd 2> /dev/null
     sudo cp /home/admin/assets/lnd.bitcoin.conf /mnt/hdd/lnd/lnd.conf
@@ -186,16 +184,13 @@ if [ "${lightning}" == "lnd" ]; then
   # check if now a config exists
   configLinkedCorrectly=$(sudo ls sudo ls /home/bitcoin/.lnd/lnd.conf | grep -c "lnd.conf")
   if [ "${configLinkedCorrectly}" != "1" ]; then
-    sed -i "s/^state=.*/state=error/g" ${infoFile}
-    sed -i "s/^message=.*/message='setup: lnd conf link broken'/g" ${infoFile}
-    echo "FAIL see ${logFile}"
-    echo "FAIL: setup: lnd conf link broken" >> ${logFile}
+    /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-link-broken" "link /home/bitcoin/.lnd/lnd.conf broken" "" ${logFile}
     exit 7
   fi
 
   # Init LND service & start
   echo "*** Init LND Service & Start ***" >> ${logFile}
-  sudo sed -i "s/^message=.*/message='LND Testrun'/g" ${infoFile}
+  /home/admin/_cache.sh set message "LND Testrun"
 
   # just in case
   sudo systemctl stop lnd 2>/dev/null
@@ -208,13 +203,14 @@ if [ "${lightning}" == "lnd" ]; then
   echo "Starting LND Service ..." >> ${logFile}
   sudo systemctl enable lnd >> ${logFile}
   sudo systemctl start lnd >> ${logFile}
+  echo "Starting LND Service ... executed" >> ${logFile}
 
   # check that lnd started
   lndRunning=0
   loopcount=0
   while [ ${lndRunning} -eq 0 ]
   do
-    lndRunning=$(sudo systemctl status lnd.service | grep -c running)
+    lndRunning=$(systemctl status lnd.service | grep -c running)
     if [ ${lndRunning} -eq 0 ]; then
       date +%s >> ${logFile}
       echo "LND not ready yet ... waiting another 60 seconds." >> ${logFile}
@@ -222,10 +218,7 @@ if [ "${lightning}" == "lnd" ]; then
     fi
     loopcount=$(($loopcount +1))
     if [ ${loopcount} -gt 100 ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: failed lnd start'/g" ${infoFile}
-      echo "FAIL see ${logFile}"
-      echo "FAIL: setup: failed lnd start" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-start-fail" "lnd service not getting to running status" "sudo systemctl status lnd.service | grep -c running --> ${lndRunning}" ${logFile}
       exit 8
     fi
   done
@@ -233,25 +226,18 @@ if [ "${lightning}" == "lnd" ]; then
   sleep 10
 
   # Check LND health/fails (to be extended)
-  tlsExists=$(sudo ls /mnt/hdd/lnd/tls.cert 2>/dev/null | grep -c "tls.cert")
+  tlsExists=$(ls /mnt/hdd/lnd/tls.cert 2>/dev/null | grep -c "tls.cert")
   if [ ${tlsExists} -eq 0 ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: missing lnd tls'/g" ${infoFile}
-      echo "FAIL see ${logFile}"
-      echo "FAIL: setup: missing lnd tls" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-no-tls" "lnd not created TLS cert" "no /mnt/hdd/lnd/tls.cert" ${logFile}
       exit 9
   fi
 
   # import static channel backup if was uploaded
   if [ "${staticchannelbackup}" != "" ]; then
     echo "Preparing static channel backup file ${staticchannelbackup} ..." >> ${logFile}
-    source <(sudo /home/admin/config.scripts/lnd.backup.sh scb-import ${staticchannelbackup})
+    source <(/home/admin/config.scripts/lnd.backup.sh scb-import ${staticchannelbackup})
     if [ "${error}" != "" ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: lnd import SCB failed'/g" ${infoFile}
-      echo "FAIL see ${logFile}"
-      echo "FAIL: setup: lnd import SCB failed" >> ${logFile}
-      echo "${error}" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-scb-import" "lnd.backup.sh scb-import returned error" "/home/admin/config.scripts/lnd.backup.sh scb-import ${staticchannelbackup} --> ${error}" ${logFile}
       exit 10
     fi
   fi
@@ -260,33 +246,23 @@ if [ "${lightning}" == "lnd" ]; then
   if [ "${seedWords}" != "" ] && [ "${staticchannelbackup}" != "" ]; then
 
     echo "WALLET --> SEED + SCB " >> ${logFile}
-    sudo sed -i "s/^message=.*/message='LND Wallet (SEED & SCB)'/g" ${infoFile}
-    if ! pip list | grep grpc; then sudo -H python3 -m pip install grpcio==1.38.1; fi  
-    sudo /home/admin/config.scripts/lnd.initwallet.py scb mainnet ${passwordC} "${seedWords}" "${staticchannelbackup}" ${seedPassword}
+    /home/admin/_cache.sh set message "LND Wallet (SEED & SCB)"
+    if ! pip list | grep grpc; then sudo -H python3 -m pip install grpcio==1.38.1; fi
+    source <(/home/admin/config.scripts/lnd.initwallet.py scb mainnet ${passwordC} "${seedWords}" "${staticchannelbackup}" ${seedPassword})
     if [ "${err}" != "" ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: lnd wallet SCB failed'/g" ${infoFile}
-      echo "FAIL see ${logFile}"
-      echo "FAIL: setup: lnd wallet SCB failed" >> ${logFile}
-      echo "${err}" >> ${logFile}
-      echo "${errMore}" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-wallet-seed+scb" "lnd.initwallet.py scb returned error" "/home/admin/config.scripts/lnd.initwallet.py scb mainnet ... --> ${err} + ${errMore}" ${logFile}
       exit 11
     fi
 
   # WALLET --> SEED
   elif [ "${seedWords}" != "" ]; then
     
-    echo "WALLET --> SEED" >> ${logFile}
-    sudo sed -i "s/^message=.*/message='LND Wallet (SEED)'/g" ${infoFile}
+    echo "WALLET --> SEED" >> ${logFile} 
+    /home/admin/_cache.sh set message "LND Wallet (SEED)"
     if ! pip list | grep grpc; then sudo -H python3 -m pip install grpcio==1.38.1; fi  
-    sudo /home/admin/config.scripts/lnd.initwallet.py seed mainnet ${passwordC} "${seedWords}" ${seedPassword}
+    source <(/home/admin/config.scripts/lnd.initwallet.py seed mainnet ${passwordC} "${seedWords}" ${seedPassword})
     if [ "${err}" != "" ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: lnd wallet SEED failed'/g" ${infoFile}
-      echo "FAIL see ${logFile}"
-      echo "FAIL: setup: lnd wallet SEED failed" >> ${logFile}
-      echo "${err}" >> ${logFile}
-      echo "${errMore}" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-wallet-seed" "lnd.initwallet.py seed returned error" "/home/admin/config.scripts/lnd.initwallet.py seed mainnet ... --> ${err} + ${errMore}" ${logFile}
       exit 12
     fi
 
@@ -294,14 +270,15 @@ if [ "${lightning}" == "lnd" ]; then
   else
 
     echo "WALLET --> NEW" >> ${logFile}
-    sudo sed -i "s/^message=.*/message='LND Wallet (NEW)'/g" ${infoFile}
-    if ! pip list | grep grpc; then sudo -H python3 -m pip install grpcio==1.38.1; fi  
-    source <(sudo /home/admin/config.scripts/lnd.initwallet.py new mainnet ${passwordC})
+    /home/admin/_cache.sh set message "LND Wallet (NEW)"
+    if ! pip list | grep grpc; then sudo -H python3 -m pip install grpcio==1.38.1; fi 
+    source <(/home/admin/config.scripts/lnd.initwallet.py new mainnet ${passwordC})
     if [ "${err}" != "" ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: lnd wallet SEED failed'/g" ${infoFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-wallet-new" "lnd.initwallet.py new returned error" "/home/admin/config.scripts/lnd.initwallet.py new mainnet ... --> ${err} + ${errMore}" ${logFile}
+      /home/admin/_cache.sh set state "error"
+      /home/admin/_cache.sh set message "setup: lnd wallet NEW failed"
       echo "FAIL see ${logFile}"
-      echo "FAIL: setup: lnd wallet SEED failed" >> ${logFile}
+      echo "FAIL: setup: lnd wallet SEED failed (2)" >> ${logFile}
       echo "${err}" >> ${logFile}
       echo "${errMore}" >> ${logFile}
       exit 13
@@ -315,14 +292,12 @@ if [ "${lightning}" == "lnd" ]; then
 
   # sync macaroons & TLS to other users
   echo "*** Copy LND Macaroons to user admin ***" >> ${logFile}
-  sudo sed -i "s/^message=.*/message='LND Credentials'/g" ${infoFile}    
+  /home/admin/_cache.sh set message "LND Credentials"
 
   # check if macaroon exists now - if not fail
   macaroonExists=$(sudo -u bitcoin ls -la /home/bitcoin/.lnd/data/chain/${network}/${chain}net/admin.macaroon 2>/dev/null | grep -c admin.macaroon)
   if [ ${macaroonExists} -eq 0 ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: lnd no macaroons'/g" ${infoFile}
-      echo "FAIL: setup: lnd no macaroons" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-no-macaroons" "lnd did not create macaroons" "/home/bitcoin/.lnd/data/chain/${network}/${chain}net/admin.macaroon --> missing" ${logFile}
       exit 14
   fi
 
@@ -332,10 +307,7 @@ if [ "${lightning}" == "lnd" ]; then
   # make a final lnd check
   source <(/home/admin/config.scripts/lnd.check.sh basic-setup)
   if [ "${err}" != "" ]; then
-    sed -i "s/^state=.*/state=error/g" ${infoFile}
-    sed -i "s/^message=.*/message='setup: lnd wallet SEED failed'/g" ${infoFile}
-    echo "FAIL: setup: lnd wallet SEED failed" >> ${logFile}
-    echo "${err}" >> ${logFile}
+    /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "lnd-check-error" "lnd.check.sh basic-setup with error" "/home/admin/config.scripts/lnd.check.sh basic-setup --> ${err}" ${logFile}
     exit 15
   fi
 
@@ -351,21 +323,17 @@ if [ "${lightning}" == "cl" ]; then
   # c-lightning
   echo "############## c-lightning" >> ${logFile}
 
-  sudo sed -i "s/^message=.*/message='C-Lightning Install'/g" ${infoFile}
+  /home/admin/_cache.sh set message "C-Lightning Install"
   sudo /home/admin/config.scripts/cl.install.sh on mainnet >> ${logFile}
-  sudo sed -i "s/^message=.*/message='C-Lightning Setup'/g" ${infoFile}
+  /home/admin/_cache.sh set message "C-Lightning Setup"
 
   # OLD WALLET FROM CLIGHTNING RESCUE
   if [ "${clrescue}" != "" ]; then
 
     echo "Restore CL data from uploaded rescue file ${clrescue} ..." >> ${logFile}
-    source <(sudo /home/admin/config.scripts/cl.backup.sh cl-import ${clrescue})
+    source <(/home/admin/config.scripts/cl.backup.sh cl-import ${clrescue})
     if [ "${error}" != "" ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: cl import backup failed'/g" ${infoFile}
-      echo "FAIL see ${logFile}"
-      echo "FAIL: setup: cl import backup failed" >> ${logFile}
-      echo "${error}" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "cl-import-backup" "cl.backup.sh cl-import with error" "/home/admin/config.scripts/cl.backup.sh cl-import ${clrescue} --> ${error}" ${logFile}
       exit 16
     fi
 
@@ -373,14 +341,12 @@ if [ "${lightning}" == "cl" ]; then
   elif [ "${seedWords}" != "" ]; then
 
     echo "Restore CL wallet from seedWords ..." >> ${logFile}
-    source <(sudo /home/admin/config.scripts/cl.hsmtool.sh seed-force mainnet "${seedWords}" "${seedPassword}")
+    source <(/home/admin/config.scripts/cl.hsmtool.sh seed-force mainnet "${seedWords}" "${seedPassword}")
 
     # check if wallet really got created 
-    walletExistsNow=$(sudo ls /home/bitcoin/.lightning/bitcoin/hsm_secret 2>/dev/null | grep -c "hsm_secret")
+    walletExistsNow=$(ls /home/bitcoin/.lightning/bitcoin/hsm_secret 2>/dev/null | grep -c "hsm_secret")
     if [ $walletExistsNow -eq 0 ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: seed maybe wrong'/g" ${infoFile}
-      echo "FAIL: setup: no cl wallet created - seed maybe wrong" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "cl-wallet-seed" "cl.hsmtool.sh seed-force not created wallet" "ls /home/bitcoin/.lightning/bitcoin/hsm_secret --> 0" ${logFile}
       exit 17
     fi
 
@@ -390,22 +356,18 @@ if [ "${lightning}" == "cl" ]; then
     echo "Generate new CL wallet ..." >> ${logFile}
 
     # generate new wallet
-    source <(sudo /home/admin/config.scripts/cl.hsmtool.sh new-force mainnet)
+    source <(/home/admin/config.scripts/cl.hsmtool.sh new-force mainnet)
 
     # check if got new seedwords
     if [ "${seedwords}" == "" ] || [ "${seedwords6x4}" == "" ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: no cl seedwords'/g" ${infoFile}
-      echo "FAIL: setup: no cl seedwords" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "cl-wallet-new-noseeds" "cl.hsmtool.sh new-force did not returned seedwords" "/home/admin/config.scripts/cl.hsmtool.sh new-force mainnet --> seedwords=''" ${logFile}
       exit 18
     fi
 
     # check if wallet really got created 
     walletExistsNow=$(sudo ls /home/bitcoin/.lightning/bitcoin/hsm_secret 2>/dev/null | grep -c "hsm_secret")
     if [ $walletExistsNow -eq 0 ]; then
-      sed -i "s/^state=.*/state=error/g" ${infoFile}
-      sed -i "s/^message=.*/message='setup: no cl wallet created'/g" ${infoFile}
-      echo "FAIL: setup: no cl wallet created" >> ${logFile}
+      /home/admin/config.scripts/blitz.error.sh _provision.setup.sh "cl-wallet-new-nowallet" "cl.hsmtool.sh new-force did not created wallet" "/home/bitcoin/.lightning/bitcoin/hsm_secret --> missing" ${logFile}
       exit 19
     fi
 
@@ -425,6 +387,6 @@ fi
 echo "stopping bitcoind for the rest provision again (will start on next boot)" >> ${logFile}
 systemctl stop bitcoind >> ${logFile}
 
-sudo sed -i "s/^message=.*/message='Provision Setup Finish'/g" ${infoFile}
+/home/admin/_cache.sh set message "Provision Setup Finish"
 echo "END Setup"  >> ${logFile}
 exit 0
